requirements.txt
mlx-whisper
pyannote.audio
python-dotenv
pydub
huggingface_hub
torch
torchvision
torchaudio
